requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.1.13
django-rest-framework==0.1.0
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.5.0
idna==3.6
kombu==5.3.5
prompt-toolkit==3.0.43
pymongo==3.12.3
python-dateutil==2.8.2
pytz==2023.4
requests==2.31.0
six==1.16.0
sqlparse==0.2.4
tzdata==2023.4
urllib3==2.2.0
vine==5.1.0
wcwidth==0.2.13
